Disable logins when installing the desktop environment. Provide support for not installing the desktop with Ubuntu 1404.

+++ ubuntu-1404/install.sh
@@ -4,6 +4,17 @@ STARTDATE=$(date)
 {
 set -x
 
+TYPES="base"
+
+if [[ $1 == desktop ]]; then
+  echo "Installation in progress, please wait" > /etc/nologin
+  # Install a full desktop (use this on Windows if you haven't installed an X server)
+  TYPES="$TYPES desktop"
+fi
+
+# Set up system to work with the Met Office Science Repository Service
+TYPES="$TYPES mosrs"
+
 # Address issues some hosts experience with networking (specifically, DNS latency)
 # See https://github.com/mitchellh/vagrant/issues/1172
 if [ ! $(grep single-request-reopen /etc/resolvconf/resolv.conf.d/base) ]; then
@@ -21,14 +32,6 @@ apt-get upgrade -y
 # Use dos2unix in case any files have Windows EOL characters
 apt-get install -y dos2unix
 
-TYPES="base"
-
-# Install a full desktop (use this on Windows if you haven't installed an X server)
-TYPES="$TYPES desktop"
-
-# Set up system to work with the Met Office Science Repository Service
-TYPES="$TYPES mosrs"
-
 for TYPE in $TYPES; do
   dos2unix -n /vagrant/install-$TYPE.sh /tmp/install-$TYPE.sh
   . /tmp/install-$TYPE.sh
@@ -41,9 +44,13 @@ apt-get autoremove -y
 set +x
 echo Finished provisioning at $(date) \(started at $STARTDATE\)
 echo
-echo Shutting down the system.
-echo Please run vagrant up to restart it.
 
+if [[ $1 == desktop ]]; then
+  echo Shutting down the system.
+  echo Please run vagrant up to restart it.
+  rm /etc/nologin
+  sudo shutdown -h now
+else
+  echo Please run vagrant ssh to connect.
+fi
 } |& tee -a /var/log/install.log
-
-sudo shutdown -h now
